Rename and move `gumtree` package
