+++ src/main/kotlin/org/jetbrains/research/plugin/modifying/PyElementTransformer.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.research.gumtree
+package org.jetbrains.research.plugin.modifying
 
 import com.github.gumtreediff.actions.model.Delete
 import com.github.gumtreediff.actions.model.Insert
