+++ src/main/kotlin/org/jetbrains/research/plugin/modifying/PyPsiGumTree.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.research.gumtree
+package org.jetbrains.research.plugin.modifying
 
 import com.github.gumtreediff.tree.ITree
 import com.github.gumtreediff.tree.Tree
